// gen/ts/open_splunk/v1/app.ts
export interface AppSelector {
  selector: { $case: "appId"; value: string } | { $case: "slug"; value: string } | undefined;
}
// ...
export const AppSelector: MessageFns<AppSelector> = {
  encode(message: AppSelector, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    switch (message.selector?.$case) {
      case "appId":
        writer.uint32(10).string(message.selector.value);
        break;
      case "slug":
        writer.uint32(18).string(message.selector.value);
        break;
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppSelector {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppSelector();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.selector = { $case: "appId", value: reader.string() };
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.selector = { $case: "slug", value: reader.string() };
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AppSelector {
    return {
      selector: isSet(object.appId)
        ? { $case: "appId", value: globalThis.String(object.appId) }
        : isSet(object.app_id)
        ? { $case: "appId", value: globalThis.String(object.app_id) }
        : isSet(object.slug)
        ? { $case: "slug", value: globalThis.String(object.slug) }
        : undefined,
    };
  },

  toJSON(message: AppSelector): unknown {
    const obj: any = {};
    if (message.selector?.$case === "appId") {
      obj.appId = message.selector.value;
    } else if (message.selector?.$case === "slug") {
      obj.slug = message.selector.value;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AppSelector>, I>>(base?: I): AppSelector {
    return AppSelector.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppSelector>, I>>(object: I): AppSelector {
    const message = createBaseAppSelector();
    switch (object.selector?.$case) {
      case "appId": {
        if (object.selector?.value !== undefined && object.selector?.value !== null) {
          message.selector = { $case: "appId", value: object.selector.value };
        }
        break;
      }
      case "slug": {
        if (object.selector?.value !== undefined && object.selector?.value !== null) {
          message.selector = { $case: "slug", value: object.selector.value };
        }
        break;
      }
    }
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string; value: unknown } ? { $case: T["$case"]; value?: DeepPartial<T["value"]> }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}