// gen/ts/open_splunk/v1/app_api.ts
export interface GetAppRequest {
  selector: AppSelector | undefined;
}
// ...
export const GetAppRequest: MessageFns<GetAppRequest> = {
  encode(message: GetAppRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.selector !== undefined) {
      AppSelector.encode(message.selector, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetAppRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetAppRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.selector = AppSelector.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetAppRequest {
    return { selector: isSet(object.selector) ? AppSelector.fromJSON(object.selector) : undefined };
  },

  toJSON(message: GetAppRequest): unknown {
    const obj: any = {};
    if (message.selector !== undefined) {
      obj.selector = AppSelector.toJSON(message.selector);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetAppRequest>, I>>(base?: I): GetAppRequest {
    return GetAppRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetAppRequest>, I>>(object: I): GetAppRequest {
    const message = createBaseGetAppRequest();
    message.selector = (object.selector !== undefined && object.selector !== null)
      ? AppSelector.fromPartial(object.selector)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string; value: unknown } ? { $case: T["$case"]; value?: DeepPartial<T["value"]> }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}